feat: move calculate_markup function to its own module and update imports
