+++ dinero/tools/__init__.py
@@ -1,7 +1,7 @@
 from .calculate_compound_interest import calculate_compound_interest
-from .calculate_markup import calculate_markup
 from .calculate_percentage import calculate_percentage
 from .calculate_simple_interest import calculate_simple_interest
+from .markup import calculate_markup
 from .vat import calculate_gross_amount, calculate_vat_portion
 
 __all__ = [
